ALL WORKING

+++ Lab06/2.ImagesHomog/include/printFigure.m
@@ -19,7 +19,7 @@ function printFigure(numberImg, numberRow, inputImg, method, subplotTitle, flag)
                 subplot(numberRow, numberImg/numberRow, i); imhist(uint8(inputImg{i}), 256); colormap gray; title(subplotTitle{i}); axis on;
                 
             case 3 % imshow
-                subplot(numberRow, numberImg/numberRow, i); imshow(inputImg{i}); colormap gray; title(subplotTitle{i}); axis on;
+                subplot(numberRow, numberImg/numberRow, i); imshow(uint8(inputImg{i})); colormap gray; title(subplotTitle{i}); axis on;
                 
             case 4 % mesh
                 subplot(numberRow, numberImg/numberRow, i); mesh(inputImg{i}); colormap gray; title(subplotTitle{i}); axis on;
